agent-service/requirements.txt
APScheduler==3.11.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
distro==1.9.0
dateparser==1.2.2
fastapi==0.115.5
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-openai==0.3.35
langfuse==2.57.13
langgraph==0.2.76
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.1
langgraph-sdk==0.1.74
langsmith==0.7.17
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
protobuf==6.33.5
psycopg2-binary==2.9.11
python-socketio==5.16.1
pydantic==2.12.5
PyJWT==2.12.1
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.41.3
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
pytest==8.4.2
